requirements.txt
streamlit
boto3
pinecone-client
openai
python-dotenv
momento
